Restructured user api. Created user directory inside endpoints and splited each route into seperate files
