+++ src/endpoints/user/router_init.py
@@ -1,0 +1,5 @@
+from fastapi import APIRouter
+
+router = APIRouter(
+    prefix="/user", tags=["user v2"], responses={401: {"user": "Not authorized"}}
+)
